move global observable to its own file
