+++ client/Tutorial/TutorialSpy.ts
@@ -1,0 +1,3 @@
+import * as ko from "knockout";
+
+export const TutorialSpy = ko.observable<string>(null);
